Decode spectrum intensity as float32, not int32
Spectrum::intensity_type and decode_intensity() used DataType::Int32, but
the intensity array is float32 (MS:1000521).  parquet_array_cast<Int32> on
a FloatArray reinterprets the underlying bytes, so every intensity value
was garbage (e.g. 1156727746 instead of 1938.12).

Use DataType::Float32 for the intensity type and decoder.

+++ include/mzpeak/spectrum.h
@@ -29,7 +29,7 @@ public:
 
   /// Ensure types stay in sync.
   using intensity_type =
-      Schema::PSI::data_type_traits<Schema::PSI::DataType::Int32>::value_type;
+      Schema::PSI::data_type_traits<Schema::PSI::DataType::Float32>::value_type;
 
   /// Destructor.
   ~Spectrum() = default;
